network: use vector<dnet_addr> instead of dnet_addr[] & move some common exception-safe guards to common.hpp
